+++ core/app_apis_administration.go
@@ -32,9 +32,10 @@ func (app *Application) getNudgesConfig(l *logs.Log) (*model.NudgesConfig, error
 	return nudgesConfig, nil
 }
 
-func (app *Application) updateNudgesConfig(l *logs.Log, active bool, groupName string, testGroupName string, mode string) error {
-	nudgesConfig := model.NudgesConfig{Active: active, GroupName: groupName, TestGroupName: testGroupName, Mode: mode}
-	err := app.storage.UpdateNudgesConfig(nudgesConfig)
+func (app *Application) updateNudgesConfig(l *logs.Log, active bool, groupName string, testGroupName string, mode string, processTime *int) error {
+	nudgesConfig := model.NudgesConfig{Active: active, GroupName: groupName, TestGroupName: testGroupName, Mode: mode, ProcessTime: processTime}
+
+	err := app.storage.SaveNudgesConfig(nudgesConfig)
 	if err != nil {
 		return err
 	}