+++ core/app_nudges.go
@@ -84,7 +84,10 @@ func (n nudgesLogic) setupNudgesTimer() {
 	n.logger.Infof("setupNudgesTimer -> now - hours:%d minutes:%d seconds:%d\n", now.Hour(), now.Minute(), now.Second())
 
 	nowSecondsInDay := 60*60*now.Hour() + 60*now.Minute() + now.Second()
-	desiredMoment := 39600 //desired moment in the day in seconds, i.e. 11:00 AM
+	desiredMoment := 39600 //default desired moment in the day in seconds, i.e. 11:00 AM
+	if n.config != nil && n.config.ProcessTime != nil {
+		desiredMoment = *n.config.ProcessTime
+	}
 
 	var durationInSeconds int
 	n.logger.Infof("setupNudgesTimer -> nowSecondsInDay:%d desiredMoment:%d\n", nowSecondsInDay, desiredMoment)
@@ -138,9 +141,8 @@ func (n nudgesLogic) processNudges() {
 	}
 }
 
-//TODO - decide if we need to loop through nudges or through all users(are the users the same for the nudges?)
 func (n nudgesLogic) processAllNudges() {
-	n.logger.Info("processAllNudges")
+	n.logger.Info("START nudges processing")
 
 	//1. first check if we have a config and the config is set to active
 	if n.config == nil {
@@ -153,28 +155,97 @@ func (n nudgesLogic) processAllNudges() {
 	}
 	n.logger.Info("the nudges processing is active")
 
-	//2. get all active nudges
-	nudges, err := n.storage.LoadActiveNudges()
-	if err != nil {
-		n.logger.Errorf("error on processing all nudges - %s", err)
-		return
-	}
-	if len(nudges) == 0 {
-		n.logger.Info("no active nudges for processing")
-	}
+	//TODO start process and keep the users for processing on pieces
 
-	//3. get all users
-	groupName := n.getGroupName()
-	users, err := n.groupsBB.GetUsers(groupName)
+	//2. process phase 1
+	err := n.processPhase1()
 	if err != nil {
-		n.logger.Errorf("error getting all users - %s", err)
+		n.logger.Errorf("error on processing phase 1 - %s", err)
 		return
 	}
 
-	//4. process every nudge
-	for _, nudge := range nudges {
-		n.processNudge(nudge, users)
+	/*
+		//4. get all active nudges
+		nudges, err := n.storage.LoadActiveNudges()
+		if err != nil {
+			n.logger.Errorf("error on processing all nudges - %s", err)
+			return
+		}
+		if len(nudges) == 0 {
+			n.logger.Info("no active nudges for processing")
+		}
+
+		//5. process every user
+		for _, user := range users {
+			n.processUser(user, nudges)
+		} */
+}
+
+//as a result of phase 1 we have into our service a cached provider data for:
+// all users
+// users courses
+// courses assignments
+// - with acceptable sync date
+func (n nudgesLogic) processPhase1() error {
+	n.logger.Info("START Phase1")
+	/*
+		//2. get all users
+		groupName := n.getGroupName()
+		users, err := n.groupsBB.GetUsers(groupName)
+		if err != nil {
+			n.logger.Errorf("error getting all users - %s", err)
+			return
+		}
+		if len(users) == 0 {
+			n.logger.Info("no users for processing")
+			return
+		}
+
+		//3. prepare(cache/optimise) the provider data
+		err = n.prepareProviderData(users)
+		if err != nil {
+			n.logger.Errorf("error on preparing the provider data - %s", err)
+			return
+		} */
+
+	n.logger.Info("END Phase1")
+	return nil
+}
+
+//phase2 operates over the data prepared in phase1 and apply the nudges for every user
+func (n nudgesLogic) processPhase2() error {
+	return nil
+}
+
+func (n nudgesLogic) prepareProviderData(users []GroupsBBUser) error {
+	n.logger.Info("\tprepareProviderData")
+
+	//get the net ids from the users
+	usersIDs := n.prepareUsers(users)
+	if len(usersIDs) == 0 {
+		n.logger.Info("\t\tno users for processing")
+		return nil
 	}
+
+	//process caching
+	err := n.provider.CacheCommonData(usersIDs)
+	if err != nil {
+		n.logger.Errorf("error caching common data- %s", err)
+		return err
+	}
+
+	return nil
+}
+
+//returns the net ids for all user who have it
+func (n nudgesLogic) prepareUsers(users []GroupsBBUser) map[string]string {
+	result := map[string]string{}
+	for _, user := range users {
+		if len(user.NetID) > 0 {
+			result[user.NetID] = user.UserID
+		}
+	}
+	return result
 }
 
 func (n nudgesLogic) getGroupName() string {
@@ -184,18 +255,24 @@ func (n nudgesLogic) getGroupName() string {
 	return n.config.TestGroupName //test mode
 }
 
-func (n nudgesLogic) processNudge(nudge model.Nudge, allUsers []GroupsBBUser) {
-	n.logger.Infof("processNudge - %s", nudge.ID)
+func (n nudgesLogic) processUser(user GroupsBBUser, nudges []model.Nudge) {
+	for _, nudge := range nudges {
+		n.processNudge(nudge, user)
+	}
+}
+
+func (n nudgesLogic) processNudge(nudge model.Nudge, user GroupsBBUser) {
+	n.logger.Infof("processNudge - %s - %s", user.NetID, nudge.ID)
 
 	switch nudge.ID {
 	case "last_login":
-		n.processLastLoginNudge(nudge, allUsers)
+		n.processLastLoginNudgePerUser(nudge, user)
 	case "missed_assignment":
-		n.processMissedAssignmentNudge(nudge, allUsers)
+		n.processMissedAssignmentNudgePerUser(nudge, user)
 	case "completed_assignment_early":
-		n.processCompletedAssignmentEarlyNudge(nudge, allUsers)
+		n.processCompletedAssignmentEarlyNudgePerUser(nudge, user)
 	case "today_calendar_events":
-		n.processTodayCalendarEventsNudge(nudge, allUsers)
+		n.processTodayCalendarEventsNudgePerUser(nudge, user)
 	default:
 		n.logger.Infof("Not supported nudge - %s", nudge.ID)
 	}
@@ -213,16 +290,8 @@ func (n nudgesLogic) prepareNotificationData(deepLink string) map[string]string 
 
 // last_login nudge
 
-func (n nudgesLogic) processLastLoginNudge(nudge model.Nudge, allUsers []GroupsBBUser) {
-	n.logger.Infof("processLastLoginNudge - %s", nudge.ID)
-
-	for _, user := range allUsers {
-		n.processLastLoginNudgePerUser(nudge, user)
-	}
-}
-
 func (n nudgesLogic) processLastLoginNudgePerUser(nudge model.Nudge, user GroupsBBUser) {
-	n.logger.Infof("processLastLoginNudgePerUser - %s", nudge.ID)
+	n.logger.Infof("\tprocessLastLoginNudgePerUser - %s", nudge.ID)
 
 	//get last login date
 	lastLogin, err := n.provider.GetLastLogin(user.NetID)
@@ -237,7 +306,7 @@ func (n nudgesLogic) processLastLoginNudgePerUser(nudge model.Nudge, user Groups
 	}
 
 	//determine if needs to send notification
-	hours := float64(nudge.Params["hours"].(int32))
+	hours := nudge.Params["hours"].(float64)
 	now := time.Now()
 	difference := now.Sub(*lastLogin) //difference between now and the last login
 	differenceInHours := difference.Hours()
@@ -307,16 +376,8 @@ func (n nudgesLogic) createSentNudge(nudgeID string, userID string, netID string
 
 // missed_assignemnt nudge
 
-func (n nudgesLogic) processMissedAssignmentNudge(nudge model.Nudge, allUsers []GroupsBBUser) {
-	n.logger.Infof("processMissedAssignmentNudge - %s", nudge.ID)
-
-	for _, user := range allUsers {
-		n.processMissedAssignmentNudgePerUser(nudge, user)
-	}
-}
-
 func (n nudgesLogic) processMissedAssignmentNudgePerUser(nudge model.Nudge, user GroupsBBUser) {
-	n.logger.Infof("processMissedAssignmentNudgePerUser - %s", nudge.ID)
+	n.logger.Infof("\tprocessMissedAssignmentNudgePerUser - %s", nudge.ID)
 
 	//get missed assignments
 	missedAssignments, err := n.provider.GetMissedAssignments(user.NetID)
@@ -330,7 +391,7 @@ func (n nudgesLogic) processMissedAssignmentNudgePerUser(nudge model.Nudge, user
 	}
 
 	//determine for which of the assignments we need to send notifications
-	hours := float64(nudge.Params["hours"].(int32))
+	hours := nudge.Params["hours"].(float64)
 	now := time.Now()
 	missedAssignments, err = n.findMissedAssignments(hours, now, missedAssignments)
 	if err != nil {
@@ -428,16 +489,8 @@ func (n nudgesLogic) findMissedAssignments(hours float64, now time.Time, assignm
 
 // completed_assignment_early nudge
 
-func (n nudgesLogic) processCompletedAssignmentEarlyNudge(nudge model.Nudge, allUsers []GroupsBBUser) {
-	n.logger.Infof("processCompletedAssignmentEarlyNudge - %s", nudge.ID)
-
-	for _, user := range allUsers {
-		n.processCompletedAssignmentEarlyNudgePerUser(nudge, user)
-	}
-}
-
 func (n nudgesLogic) processCompletedAssignmentEarlyNudgePerUser(nudge model.Nudge, user GroupsBBUser) {
-	n.logger.Infof("processCompletedAssignmentEarlyNudgePerUser - %s", nudge.ID)
+	n.logger.Infof("\tprocessCompletedAssignmentEarlyNudgePerUser - %s", nudge.ID)
 
 	//get completed assignments
 	ecAssignments, err := n.provider.GetCompletedAssignments(user.NetID)
@@ -451,7 +504,7 @@ func (n nudgesLogic) processCompletedAssignmentEarlyNudgePerUser(nudge model.Nud
 	}
 
 	//determine for which of the submissions we need to send notifications
-	hours := float64(nudge.Params["hours"].(int32))
+	hours := nudge.Params["hours"].(float64)
 	now := time.Now()
 	ecAssignments, err = n.findCompletedEarlyAssignments(hours, now, ecAssignments)
 	if err != nil {
@@ -551,16 +604,9 @@ func (n nudgesLogic) sendEarlyCompletedAssignmentNudgeForUser(nudge model.Nudge,
 // end completed_assignment_early nudge
 
 // calendar_event nudge
-func (n nudgesLogic) processTodayCalendarEventsNudge(nudge model.Nudge, allUsers []GroupsBBUser) {
-	n.logger.Infof("processTodayCalendarEventsNudge - %s", nudge.ID)
-
-	for _, user := range allUsers {
-		n.processTodayCalendarEventsNudgePerUser(nudge, user)
-	}
-}
 
 func (n nudgesLogic) processTodayCalendarEventsNudgePerUser(nudge model.Nudge, user GroupsBBUser) {
-	n.logger.Infof("processTodayCalendarEventsNudgePerUser - %s", nudge.ID)
+	n.logger.Infof("\tprocessTodayCalendarEventsNudgePerUser - %s", nudge.ID)
 
 	//get calendar events
 	startDate, endDate := n.prepareTodayCalendarEventsDates()