+++ core/interfaces.go
@@ -36,7 +36,7 @@ type Services interface {
 // Administration exposes APIs for the driver adapters
 type Administration interface {
 	GetNudgesConfig(l *logs.Log) (*model.NudgesConfig, error)
-	UpdateNudgesConfig(l *logs.Log, active bool, groupName string, testGroupName string, mode string) error
+	UpdateNudgesConfig(l *logs.Log, active bool, groupName string, testGroupName string, mode string, processTime *int) error
 
 	GetNudges() ([]model.Nudge, error)
 	CreateNudge(l *logs.Log, ID string, name string, body string, deepLink string, params *map[string]interface{}, active bool) error
@@ -85,8 +85,8 @@ func (s *administrationImpl) GetNudgesConfig(l *logs.Log) (*model.NudgesConfig, 
 	return s.app.getNudgesConfig(l)
 }
 
-func (s *administrationImpl) UpdateNudgesConfig(l *logs.Log, active bool, groupName string, testGroupName string, mode string) error {
-	return s.app.updateNudgesConfig(l, active, groupName, testGroupName, mode)
+func (s *administrationImpl) UpdateNudgesConfig(l *logs.Log, active bool, groupName string, testGroupName string, mode string, processTime *int) error {
+	return s.app.updateNudgesConfig(l, active, groupName, testGroupName, mode, processTime)
 }
 
 func (s *administrationImpl) GetNudges() ([]model.Nudge, error) {
@@ -119,7 +119,7 @@ type Storage interface {
 
 	CreateNudgesConfig(nudgesConfig model.NudgesConfig) error
 	FindNudgesConfig() (*model.NudgesConfig, error)
-	UpdateNudgesConfig(nudgesConfig model.NudgesConfig) error
+	SaveNudgesConfig(nudgesConfig model.NudgesConfig) error
 
 	LoadAllNudges() ([]model.Nudge, error)
 	LoadActiveNudges() ([]model.Nudge, error)
@@ -141,6 +141,9 @@ type Provider interface {
 	GetAssignmentGroups(userID string, courseID int, include *string) ([]model.AssignmentGroup, error)
 	GetCourseUser(userID string, courseID int, includeEnrolments bool, includeScores bool) (*model.User, error)
 	GetCurrentUser(userID string) (*model.User, error)
+
+	CacheCommonData(usersIDs map[string]string) error
+
 	GetLastLogin(userID string) (*time.Time, error)
 	GetMissedAssignments(userID string) ([]model.Assignment, error)
 	GetCompletedAssignments(userID string) ([]model.Assignment, error)
@@ -149,7 +152,7 @@ type Provider interface {
 
 //GroupsBB interface for the Groups building block communication
 type GroupsBB interface {
-	GetUsers(groupName string) ([]GroupsBBUser, error)
+	GetUsers(groupName string, offset int, limit int) ([]GroupsBBUser, error)
 }
 
 //GroupsBBUser entity