+++ driven/provider/entities.go
@@ -26,5 +26,29 @@ type providerUser struct {
 	ID       string     `bson:"_id"`    //core BB account id
 	NetID    string     `bson:"net_id"` //core BB external system id
 	User     model.User `bson:"user"`
-	SyncDate time.Time  `json:"sync_date"`
+	SyncDate time.Time  `bson:"sync_date"`
+
+	Courses *userCourses `bson:"courses"`
+}
+
+type userCourses struct {
+	Data     []userCourse `bson:"data"`
+	SyncDate time.Time    `bson:"sync_date"`
+}
+
+type userCourse struct {
+	Data        model.Course       `bson:"data"`
+	Assignments []courseAssignment `bson:"assignments"`
+	SyncDate    time.Time          `bson:"sync_date"`
+}
+
+type courseAssignment struct {
+	Data       model.Assignment `bson:"data"`
+	Submission *submission      `bson:"submission"`
+	SyncDate   time.Time        `bson:"sync_date"`
+}
+
+type submission struct {
+	Data     model.Submission `bson:"data"`
+	SyncDate time.Time        `bson:"sync_date"`
 }