+++ driver/web/adapter.go
@@ -99,14 +99,12 @@ func (we Adapter) Start() {
 	///admin ///
 	adminRouter := subrouter.PathPrefix("/admin").Subrouter()
 
-	adminRouter.HandleFunc("/nudges-config", we.adminAuthWrapFunc(we.adminApisHandler.GetNudgesConfig)).Methods("GET")
-	adminRouter.HandleFunc("/nudges-config", we.adminAuthWrapFunc(we.adminApisHandler.UpdateNudgesConfig)).Methods("PUT")
 	adminRouter.HandleFunc("/nudges", we.adminAuthWrapFunc(we.adminApisHandler.GetNudges)).Methods("GET")
 	adminRouter.HandleFunc("/nudges", we.adminAuthWrapFunc(we.adminApisHandler.CreateNudge)).Methods("POST")
 	adminRouter.HandleFunc("/nudges/{id}", we.adminAuthWrapFunc(we.adminApisHandler.UpdateNudge)).Methods("PUT")
 	adminRouter.HandleFunc("/nudges/{id}", we.adminAuthWrapFunc(we.adminApisHandler.DeleteNudge)).Methods("DELETE")
-	adminRouter.HandleFunc("/sent-nudges", we.adminAuthWrapFunc(we.adminApisHandler.FindSentNudges)).Methods("GET")
-	adminRouter.HandleFunc("/sent-nudges", we.adminAuthWrapFunc(we.adminApisHandler.DeleteSentNudges)).Methods("DELETE")
+	adminRouter.HandleFunc("/sent_nudges", we.adminAuthWrapFunc(we.adminApisHandler.FindSentNudges)).Methods("GET")
+	adminRouter.HandleFunc("/sent_nudges", we.adminAuthWrapFunc(we.adminApisHandler.DeleteSentNudges)).Methods("")
 
 	log.Fatal(http.ListenAndServe(":"+we.port, router))
 }